refactor: Restructor folder/file tree
